Improve performance of solving over apply expressions by directly applying the return type whenever possible. This has some nice side-effects: - Addresses many common user-reported "expression too complex" bugs, including rdar://problem/18876786. - Shaves up to 10% off of the total time to run our unit tests. (Unscientifically measured on my iMac: 427.46s before, 385.17s after.)
Swift SVN r24514

+++ test/SILPasses/diagnostic_constant_propagation.swift
@@ -29,7 +29,8 @@ func testArithmeticOverflow() {
   xu8_3 += 40 // expected-error {{arithmetic operation '240 + 40' (on type 'UInt8') results in an overflow}}
   var cadd : UInt8 = 240 + 5 + 15 // expected-error {{arithmetic operation '245 + 15' (on type 'UInt8') results in an overflow}}
   
-  let ranges = Int8(126) + (1+1) // expected-error {{arithmetic operation '126 + 2' (on type 'Int8') results in an overflow}}  
+  // TODO: We should remove the second init for Int8 - see rdar://problem/19224768
+  let ranges = Int8(126) + Int8(1+1) // expected-error {{arithmetic operation '126 + 2' (on type 'Int8') results in an overflow}}  
 //  DISABLED FOR NOW
 //  asserts in the shift operators confuse constant propagation
 //  var csh1: Int8 = (1 << 7) - 1 // expected - error {{arithmetic operation '-128 - 1' (on type 'Int8') results in an overflow}}
